refactor: Move renderInfoItem to separate component
