+++ src/components/infoItem.ts
@@ -1,0 +1,57 @@
+import { html } from 'lit-element'
+
+interface InfoItemDetails {
+  heading?: string
+  icon?: string
+  unit?: string
+}
+
+interface InfoItemOptions {
+  hide: boolean
+  state: any
+  details: InfoItemDetails
+}
+
+// Preset mode can be  one of: none, eco, away, boost, comfort, home, sleep, activity
+// See https://github.com/home-assistant/home-assistant/blob/dev/homeassistant/components/climate/const.py#L36-L57
+
+export default function renderInfoItem({
+  hide,
+  state,
+  details,
+}: InfoItemOptions) {
+  if (hide || !state) return
+
+  const { heading, icon, unit } = details
+
+  let valueCell
+  if (typeof state === 'object') {
+    let value = state.state
+    if ('device_class' in state.attributes) {
+      const [type] = state.entity_id.split('.')
+      const prefix = ['state', type, state.attributes.device_class, ''].join(
+        '.'
+      )
+      value = this.localize(state.state, prefix)
+    }
+    valueCell = html`
+      <div
+        class="sensor-value clickable"
+        @click="${() => this.openEntityPopover(state.entity_id)}"
+      >
+        ${value} ${unit || state.attributes.unit_of_measurement}
+      </div>
+    `
+  } else {
+    valueCell = html` <div class="sensor-value">${state}</div> `
+  }
+
+  const headingResult = icon
+    ? html` <ha-icon icon="${icon}"></ha-icon> `
+    : html` ${heading}: `
+
+  return html`
+    <div class="sensor-heading">${headingResult}</div>
+    ${valueCell}
+  `
+}
